Rename embed-mode to amp-story-embed-mode
