+++ extensions/amp-story/1.0/amp-story-store-service.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {EmbedMode, parseEmbedMode} from '../../../src/embed-mode';
+import {EmbedMode, parseEmbedMode} from '../../../src/amp-story-embed-mode';
 import {Observable} from '../../../src/observable';
 import {Services} from '../../../src/services';
 import {deepEquals} from '../../../src/json';
