+++ src/amp-story-embed-mode.js
